authorization added

+++ EventReservation/EventReservation/Controllers/LocalsController.cs
@@ -35,124 +35,142 @@ namespace EventReservation.Controllers
         // GET: Locals/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (!User.IsInRole("Manager"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Local local = db.Locals.Find(id);
-            if (local == null)
-            {
-                return HttpNotFound();
-            }
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Local local = db.Locals.Find(id);
+                if (local == null)
+                {
+                    return HttpNotFound();
+                }
 
-            int stars = 0;
-            var result = db.Locals.Include(l => l.Reviews).FirstOrDefault(l => l.Id == local.Id).Reviews;
-            if (result.Any())
-                stars = (int)System.Math.Round(db.Locals.Include(l => l.Reviews).FirstOrDefault(l => l.Id == local.Id).Reviews.Average(review => review.Stars));
+                int stars = 0;
+                var result = db.Locals.Include(l => l.Reviews).FirstOrDefault(l => l.Id == local.Id).Reviews;
+                if (result.Any())
+                    stars = (int)System.Math.Round(db.Locals.Include(l => l.Reviews).FirstOrDefault(l => l.Id == local.Id).Reviews.Average(review => review.Stars));
 
-            int yourReview = 0;
-            var yourReviewResult = db.Reviews.FirstOrDefault(r => r.Local.Id == local.Id);
-            // && r.User == User.Identity.Name
-            if (yourReviewResult != null)
-                yourReview = db.Reviews.FirstOrDefault(r => r.Local.Id == local.Id).Stars;
-            // && r.User == User.Identity.Name
-            ViewBag.average = stars;
-            ViewBag.yourReview = yourReview;
-            return View(local);
+                int yourReview = 0;
+                var yourReviewResult = db.Reviews.FirstOrDefault(r => r.Local.Id == local.Id && r.User == User.Identity.Name);
+                if (yourReviewResult != null)
+                    yourReview = db.Reviews.FirstOrDefault(r => r.Local.Id == local.Id && r.User == User.Identity.Name).Stars;
+                ViewBag.average = stars;
+                ViewBag.yourReview = yourReview;
+                return View(local);
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         // GET: Locals/Requests
         public ActionResult Requests()
         {
-            return View("~/Views/Admin/LocalRequest.cshtml", db.LocalRequests.ToList());
+            if (User.IsInRole("Admin"))
+            {
+                return View("~/Views/Admin/LocalRequest.cshtml", db.LocalRequests.ToList());
+            }
+            return HttpNotFound();
         }
 
         //GET: Locals/Confirm/5
         public ActionResult Confirm(int id)
         {
-            //find local request
-            LocalRequest request = db.LocalRequests.Find(id);
-
-            //create and save local
-            Local local = new Local();
-            local.City = request.LocalCity;
-            local.Name = request.LocalName;
-            local.Manager = request.Email;
-            db.Locals.Add(local);
-
-            // create new user
-            var roleManager = new RoleManager<IdentityRole>(new RoleStore<IdentityRole>(db));
-            var UserManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
-            UserManager.UserValidator = new UserValidator<ApplicationUser>(UserManager)
+            if (User.IsInRole("Admin"))
             {
-                AllowOnlyAlphanumericUserNames = false,
-                RequireUniqueEmail = true
-            };
-            string password = System.Web.Security.Membership.GeneratePassword(12, 4);
-            var user = new ApplicationUser();
-            user.Email = request.Email;
-            user.UserName = request.Email;
-            var result = UserManager.Create(user, password);
-            //P@ssw0rdPassword
-            MailMessage mm = new MailMessage("eventreservationit@gmail.com", user.Email);
-            mm.Subject = "Local accepted";
-            mm.Body = "Dear " + user.Email + ", your local has been added to our webside. Thank you for choosing us. You can now loging to" +
-                "your account";
-            mm.Body += "Username: " + user.Email + "\n Password: " + password;
-            mm.IsBodyHtml = false;
+                //find local request
+                LocalRequest request = db.LocalRequests.Find(id);
 
-            SmtpClient smtp = new SmtpClient();
-            smtp.Host = "smtp.gmail.com";
-            smtp.Port = 587;
-            // smtp.EnableSsl = true;
-            NetworkCredential nc = new NetworkCredential("eventreservationit@gmail.com", "P@ssw0rdPassword");
-            smtp.UseDefaultCredentials = false;
-            smtp.Credentials = nc;
-            smtp.Send(mm);
+                //create and save local
+                Local local = new Local();
+                local.City = request.LocalCity;
+                local.Name = request.LocalName;
+                local.Manager = request.Email;
+                db.Locals.Add(local);
 
-            //add role 'Menager' to user
-            UserManager.AddToRole(user.Id, "Manager");
+                // create new user
+                var roleManager = new RoleManager<IdentityRole>(new RoleStore<IdentityRole>(db));
+                var UserManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+                UserManager.UserValidator = new UserValidator<ApplicationUser>(UserManager)
+                {
+                    AllowOnlyAlphanumericUserNames = false,
+                    RequireUniqueEmail = true
+                };
+                string password = System.Web.Security.Membership.GeneratePassword(12, 4);
+                var user = new ApplicationUser();
+                user.Email = request.Email;
+                user.UserName = request.Email;
+                var result = UserManager.Create(user, password);
+                //P@ssw0rdPassword
+                MailMessage mm = new MailMessage("eventreservationit@gmail.com", user.Email);
+                mm.Subject = "Local accepted";
+                mm.Body = "Dear " + user.Email + ", your local has been added to our webside. Thank you for choosing us. You can now loging to" +
+                    "your account";
+                mm.Body += "Username: " + user.Email + "\n Password: " + password;
+                mm.IsBodyHtml = false;
 
-            //remove request
-            db.LocalRequests.Remove(request);
-            db.SaveChanges();
-            //this needs to be made with ajax call so that it will only be deleted from the table
-            return RedirectToAction("Index", "Home");
+                SmtpClient smtp = new SmtpClient();
+                smtp.Host = "smtp.gmail.com";
+                smtp.Port = 587;
+                // smtp.EnableSsl = true;
+                NetworkCredential nc = new NetworkCredential("eventreservationit@gmail.com", "P@ssw0rdPassword");
+                smtp.UseDefaultCredentials = false;
+                smtp.Credentials = nc;
+                smtp.Send(mm);
+
+                //add role 'Menager' to user
+                UserManager.AddToRole(user.Id, "Manager");
+
+                //remove request
+                db.LocalRequests.Remove(request);
+                db.SaveChanges();
+                //this needs to be made with ajax call so that it will only be deleted from the table
+                return RedirectToAction("Index", "Home");
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         //POST: Locals/DeleteRequest
         [HttpPost]
         public ActionResult DeleteRequest(FormCollection data)
         {
-            int id = int.Parse(data["id"]);
-            //find local request
-            LocalRequest request = db.LocalRequests.Find(id);
+            if (User.IsInRole("Manager"))
+            {
+                int id = int.Parse(data["id"]);
+                //find local request
+                LocalRequest request = db.LocalRequests.Find(id);
 
-            if(request == null)
-            {
-                return HttpNotFound();
+                if (request == null)
+                {
+                    return HttpNotFound();
+                }
+                else
+                {
+                    db.LocalRequests.Remove(request);
+                    db.SaveChanges();
+                    return Json("{}");
+                }
             }
-            else
-            {
-                db.LocalRequests.Remove(request);
-                db.SaveChanges();
-                return Json("{}");
-            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         // GET: Locals/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (User.IsInRole("Manager"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Local local = db.Locals.Find(id);
+                if (local == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(local);
             }
-            Local local = db.Locals.Find(id);
-            if (local == null)
-            {
-                return HttpNotFound();
-            }
-            return View(local);
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         // POST: Locals/Edit/5
@@ -162,34 +180,38 @@ namespace EventReservation.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Name,Description,City,StreetName,StreetNo,OpeningHour,Manager,ClosingHour,Parking")] Local local)
         {
-            var files = Request.Files;
-
-            if (ModelState.IsValid)
+            if (User.IsInRole("Manager"))
             {
-                if (files != null)
-                {
-                    foreach (HttpPostedFileBase file in files)
-                    {
-                        using (var ms = new MemoryStream())
-                        {
-                            file.InputStream.CopyTo(ms);
-                            var localImage = new LocalImage
-                            {
-                                Local = local,
-                                Image = ms.ToArray()
-                            };
+                var files = Request.Files;
 
-                            local.LocalImages.Add(localImage);
+                if (ModelState.IsValid)
+                {
+                    if (files != null)
+                    {
+                        foreach (HttpPostedFileBase file in files)
+                        {
+                            using (var ms = new MemoryStream())
+                            {
+                                file.InputStream.CopyTo(ms);
+                                var localImage = new LocalImage
+                                {
+                                    Local = local,
+                                    Image = ms.ToArray()
+                                };
+
+                                local.LocalImages.Add(localImage);
+                            }
                         }
                     }
-                }
-                db.Locals.Add(local);
+                    db.Locals.Add(local);
 
-                db.Entry(local).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                    db.Entry(local).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(local);
             }
-            return View(local);
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         // POST: Locals/Delete/5
@@ -213,46 +235,63 @@ namespace EventReservation.Controllers
         }
         [HttpGet]
         public ActionResult ListEvents() {
-
-            var currentUser = User.Identity.Name;
-            var events = db.Locals
-                    .Include(l => l.Events)
-                    .Where(l => l.Manager == currentUser)
-                    .SelectMany(l => l.Events);
-            return View(events);
+            if (User.IsInRole("Manager"))
+            {
+                var currentUser = User.Identity.Name;
+                var events = db.Locals
+                        .Include(l => l.Events)
+                        .Where(l => l.Manager == currentUser)
+                        .SelectMany(l => l.Events);
+                return View(events);
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
+
+        [HttpGet]
+        public ActionResult MyLocal()
+        {
+            if (User.IsInRole("Manager"))
+            {
+                var currentUser = User.Identity.Name;
+                var local = db.Locals.
+                            Where(l => l.Manager == currentUser).
+                            FirstOrDefault();
+
+                return View(local);
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+        }
+
 
         [HttpPost]
         public ActionResult Rating(FormCollection data)
         {
-            //if (User.Identity.IsAuthenticated)
-            //{
+            double av = 0;
+            if (User.Identity.IsAuthenticated)
+            {
             int numberStars = int.Parse(data["Stars"]);
             int LocalId = int.Parse(data["Local"]);
-            var tableRow = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId);
+            var tableRow = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId && r.User == User.Identity.Name);
             int oldReview = 0;
-            // && r.User == User.Identity.Name
+            
             if (tableRow != null)
             {
-                oldReview = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId).Stars;
-                // && r.User == User.Identity.Name
+                oldReview = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId && r.User == User.Identity.Name).Stars;
                 db.Reviews.Remove(tableRow);
             }
             Local local = db.Locals.Include(l => l.Reviews).FirstOrDefault(l => l.Id == LocalId);
-            Review review = new Review { Stars = numberStars, Local = local };
-            // , User = User.Identity.Name
+            Review review = new Review { Stars = numberStars, Local = local, User = User.Identity.Name };
+ 
             db.Reviews.Add(review);
             local.Reviews.Add(review);
             if (oldReview != 0)
             {
-                var removeReview = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId);
-                // && r.User == User.Identity.Name
+                var removeReview = db.Reviews.FirstOrDefault(r => r.Local.Id == LocalId && r.User == User.Identity.Name);
                 local.Reviews.Remove(removeReview);
             }
             db.SaveChanges();
-            double av = local.Reviews.Average(r => r.Stars);
-            //}
-            //return Json(new { success = true});
+            av = local.Reviews.Average(r => r.Stars);
+            }
             return Json(new { average = av });
         }
 
@@ -260,20 +299,25 @@ namespace EventReservation.Controllers
         [HttpPost]
         public ActionResult DeleteEvent(FormCollection data)
         {
-            int id = int.Parse(data["id"]);
-            
-            Event myEvent = db.Events.Find(id);
-
-            if (myEvent == null)
+            if (User.IsInRole("Manager"))
             {
-                return HttpNotFound();
+                int id = int.Parse(data["id"]);
+
+                Event myEvent = db.Events.Find(id);
+
+                if (myEvent == null)
+                {
+                    return HttpNotFound();
+                }
+                else
+                {
+                    db.Events.Remove(myEvent);
+                    db.SaveChanges();
+                    return Json("{}");
+                }
             }
             else
-            {
-                db.Events.Remove(myEvent);
-                db.SaveChanges();
-                return Json("{}");
-            }
+                return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
     }
 }
